+++ pkg/service/config_test.go
@@ -1,11 +1,11 @@
-package identity_test
+package service_test
 
 import (
 	"testing"
 
 	configpkg "github.com/moov-io/identity/pkg/config"
-	. "github.com/moov-io/identity/pkg/identity"
 	log "github.com/moov-io/identity/pkg/logging"
+	"github.com/moov-io/identity/pkg/service"
 	"github.com/stretchr/testify/require"
 )
 
@@ -14,7 +14,7 @@ func Test_ConfigLoading(t *testing.T) {
 
 	ConfigService := configpkg.NewConfigService(logger)
 
-	config := &GlobalConfig{}
+	config := &service.GlobalConfig{}
 	err := ConfigService.Load(config)
 	require.Nil(t, err)
 }
