+++ pkg/service/environment.go
@@ -1,4 +1,4 @@
-package identity
+package service
 
 import (
 	"context"
