+++ pkg/service/environment_test.go
@@ -1,12 +1,12 @@
-package identity_test
+package service_test
 
 import (
 	"os"
 	"testing"
 
 	"github.com/go-kit/kit/log"
-	"github.com/moov-io/identity/pkg/identity"
 	"github.com/moov-io/identity/pkg/logging"
+	"github.com/moov-io/identity/pkg/service"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -14,7 +14,7 @@ func Test_Environment_Startup(t *testing.T) {
 	a := assert.New(t)
 
 	logger := logging.NewLogger(log.NewLogfmtLogger(log.NewSyncWriter(os.Stderr)))
-	env, err := identity.NewEnvironment(logger, nil)
+	env, err := service.NewEnvironment(logger, nil)
 	a.Nil(err)
 
 	shutdown := env.RunServers(false)
