+++ pkg/service/server.go
@@ -1,4 +1,4 @@
-package identity
+package service
 
 import (
 	"context"
